scripts/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
dataclasses-json==0.5.7
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
langchain==0.0.168
loguru==0.7.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mutagen==1.46.0
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
packaging==23.1
pinecone-client==2.2.1
pycryptodomex==3.17
pydantic==1.10.7
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
six==1.16.0
SQLAlchemy==2.0.13
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==2.0.2
websockets==11.0.3
yarl==1.9.2
youtube-dl==2021.12.17
yt-dlp==2023.3.4
youtube-transcript-api
